requirements.txt
flyai
Pillow
opencv-python==4.2.0.32
numpy
matplotlib
scipy
torch
torchvision
efficientnet_pytorch
resnest
cnn_finetune